Parameter B gives wrong error?
